+++ collect_app/src/main/res/layout/delete_form_layout.xml
@@ -6,7 +6,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/multi_select_list"
-        android:name="org.odk.collect.lists.multiselect.MultiSelectListFragment"
+        android:name="org.odk.collect.lists.selects.MultiSelectListFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
