--- lists/src/main/java/org/odk/collect/lists/multiselect/MultiSelectItem.kt
@@ -1,3 +1,0 @@
-package org.odk.collect.lists.multiselect
-
-data class MultiSelectItem<T>(val id: String, val item: T)
