+++ lists/src/main/java/org/odk/collect/lists/selects/MultiSelectAdapter.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.lists.multiselect
+package org.odk.collect.lists.selects
 
 import android.view.View
 import android.view.ViewGroup
@@ -19,7 +19,7 @@ class MultiSelectAdapter<T, VH : MultiSelectAdapter.ViewHolder<T>>(
             notifyDataSetChanged()
         }
 
-    var data = emptyList<MultiSelectItem<T>>()
+    var data = emptyList<SelectItem<T>>()
         set(value) {
             field = value.toList()
             notifyDataSetChanged()
