+++ lists/src/main/java/org/odk/collect/lists/selects/MultiSelectListFragment.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.lists.multiselect
+package org.odk.collect.lists.selects
 
 import android.content.Context
 import android.os.Bundle
