+++ lists/src/main/java/org/odk/collect/lists/selects/SelectItem.kt
@@ -1,0 +1,3 @@
+package org.odk.collect.lists.selects
+
+data class SelectItem<T>(val id: String, val item: T)
